add resource_sharing functions to utils
